[CLEANUP] Resolve the `BagHelper` testing trait
Part of #5095
