+++ Tests/Functional/BagBuilder/AbstractBagBuilderTest.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace OliverKlee\Seminars\Tests\Functional\BagBuilder;
 
+use OliverKlee\Seminars\Bag\AbstractBag;
 use OliverKlee\Seminars\Tests\Functional\BagBuilder\Fixtures\TestingBagBuilder;
-use OliverKlee\Seminars\Tests\Functional\Traits\BagHelper;
 use TYPO3\TestingFramework\Core\Functional\FunctionalTestCase;
 
 /**
@@ -13,8 +13,6 @@ use TYPO3\TestingFramework\Core\Functional\FunctionalTestCase;
  */
 final class AbstractBagBuilderTest extends FunctionalTestCase
 {
-    use BagHelper;
-
     protected array $testExtensionsToLoad = [
         'oliverklee/feuserextrafields',
         'oliverklee/oelib',
@@ -28,6 +26,30 @@ final class AbstractBagBuilderTest extends FunctionalTestCase
         parent::setUp();
 
         $this->subject = new TestingBagBuilder();
+    }
+
+    private static function assertBagHasUid(AbstractBag $bag, int $uid): void
+    {
+        self::assertTrue(self::bagHasUid($bag, $uid), 'The bag does not have this UID: ' . $uid);
+    }
+
+    private static function assertBagNotHasUid(AbstractBag $bag, int $uid): void
+    {
+        self::assertFalse(self::bagHasUid($bag, $uid), 'The bag has this UID, but was expected not to: ' . $uid);
+    }
+
+    private static function bagHasUid(AbstractBag $bag, int $uid): bool
+    {
+        $found = false;
+
+        foreach ($bag as $element) {
+            if ($element->getUid() === $uid) {
+                $found = true;
+                break;
+            }
+        }
+
+        return $found;
     }
 
     /**
